requirements.txt
streamlit
streamlit-autorefresh
vnstock>=3.5.0
pandas
numpy
requests
beautifulsoup4
vnai
vnstock_ezchart
flask
flask-cors
matplotlib
seaborn
openpyxl
wordcloud
mysql-connector-python
